test move function in scripts.js
